# ARTIQ Pull Request

## Description of Changes
Optimized performance for RPC list and array transfer of primitive types. Added some more tests for performance of RPC.

Tested on kc705 and zynq platform with `test_performance`, `test_embedding` and `test_numpy`. No regression.

### Notes
- Lists/arrays with non-primitive types (not Bool/Int32/Int64/Float) are not optimized that well and are in general hard to optimize well.
- This patch changes the RPC protocol for list and array, users should ensure they have the same firmware and compiler version.
- The performance test would no longer assert the data correctness. Use `test_embedding` and `test_numpy` for correctness check.
- The performance test now uses normal RPC instead of async RPC now, as this could better represent the actual throughput, and async RPC cannot be used to test for performance in artiq-zynq port. The performance number could a lot smaller than the tests before, but should be reflect the real throughput.

### RPC Performance
| Type | KC705 (previously) | KC705 (now) | Zynq (previously) | Zynq (now) |
| --- | --- | --- | --- | --- |
| test_device_to_host (bytes) | 2.0989389396479927 MiB/s | 2.099773561838202 MiB/s | 43.09186402640445 MiB/s | 43.8887923867646 MiB/s |
|  test_device_to_host_array (int32 array) | 0.19622426586254357 MiB/s | 0.4754803820141763 MiB/s | 0.4794607308006433 MiB/s | 31.9846882900218 MiB/s |
| test_device_to_host_list (int32 list) | 0.19756234991899502 MiB/s | 0.47681566599308245 MiB/s | 0.7028557219158196 MiB/s | 30.071787370811602 MiB/s |
| test_host_to_device (bytes) | 2.095725556708547 MiB/s | 2.121632685383121 MiB/s | 44.71234780449521 MiB/s | 48.77840068856528 MiB/s |
| test_host_to_device_array (int32 array) | 0.3525974608153646 MiB/s | 1.9586889006584327 MiB/s  | 0.4533021473060521 MiB/s | 21.507019202843193 MiB/s |
| test_host_to_device_byte_list (bool list) | 0.09882486389808119 MiB/s | 1.8457782297161671 MiB/s | 0.11275899857943146 MiB/s | 20.907261409677954 MiB/s |
| test_host_to_device_list (int32 list) | 0.34771873366556477 MiB/s | 1.9939231602629297 MiB/s | 0.42312216824642657 MiB/s | 21.732297044685772 MiB/s |

* The performance for `device_to_host` list/array can be optimized further in targets with allocation enabled, but this is currently only done for zynq. The performance is poor because we are calling the write function for every single element.
* There should be no significant difference in throughput for byte array, as the original implementation is already very efficient.
* The throughput of `host_to_device` list/array for zynq can be further increased to somewhere around 30MiB/s if we increase the data length, but that is just difference in numbers so I prefer to use the same length for all transactions.

## Type of Changes

<!-- Leave ONLY the corresponding lines for the applicable type of change: -->
|   | Type |
| ------------- | ------------- |
| ✓  | :hammer: Refactoring  |

## Steps (Choose relevant, delete irrelevant before submitting)

### All Pull Requests

- [x] Use correct spelling and grammar.
- [x] Update [RELEASE_NOTES.md](../RELEASE_NOTES.md) if there are noteworthy changes, especially if there are changes to existing APIs.
- [ ] Close/update issues.

### Code Changes

- [x] Run `flake8` to check code style (follow PEP-8 style). `flake8` has issues with parsing Migen/gateware code, ignore as necessary.
- [x] Test your changes or have someone test them. Mention what was tested and how.
- [ ] Add and check docstrings and comments
- [x] Check, test, and update the [unittests in /artiq/test/](../artiq/test/) or [gateware simulations in /artiq/gateware/test](../artiq/gateware/test)

### Documentation Changes

- [ ] Check, test, and update the documentation in [doc/](../doc/). Build documentation (`cd doc/manual/; make html`) to ensure no errors.

### Git Logistics

- [x] Split your contribution into logically separate changes (`git rebase --interactive`). Merge/squash/fixup commits that just fix or amend previous commits. Remove unintended changes & cleanup. See [tutorial](https://www.atlassian.com/git/tutorials/rewriting-history/git-rebase).
- [x] Write short & meaningful commit messages. Review each commit for messages (`git show`). Format:
  ```
  topic: description. < 50 characters total.
  
  Longer description. < 70 characters per line
  ```

### Licensing

See [copyright & licensing for more info](https://github.com/m-labs/artiq/blob/master/CONTRIBUTING.rst#copyright-and-sign-off).
ARTIQ files that do not contain a license header are copyrighted by M-Labs Limited and are licensed under LGPLv3+.